+++ bld/f77/wfc/c/utility.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2019 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -104,13 +104,13 @@ char    *MkNodeStr( itnode *itptr ) {
 
 // Put a NULLCHAR at the end of the "opnd" string of an itnode.
 
-    uint        opnd_size;
+    size_t      opnd_size;
     char        *str;
 
     opnd_size = itptr->opnd_size;
     str = FMemAlloc( opnd_size + sizeof( char ) );
     memcpy( str, itptr->opnd, opnd_size );
-    str[ opnd_size ] = NULLCHAR;
+    str[opnd_size] = NULLCHAR;
     return( str );
 }
 
@@ -129,7 +129,7 @@ bool    CmpNode2Str( itnode *itptr, char *str ) {
 
 // Compare the "opnd" field of an itnode to a string.
 
-    uint        str_len;
+    size_t      str_len;
 
     str_len = strlen( str );
     if( itptr->opnd_size != str_len )
