+++ bld/f77/wfc/h/usconst.h
@@ -2,6 +2,7 @@
 *
 *                            Open Watcom Project
 *
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -29,4 +30,4 @@
 ****************************************************************************/
 
 
-extern void     ConstCat( uint size );
+extern void     ConstCat( size_t size );
